+++ .nycrc
@@ -1,3 +1,6 @@
 {
-  "extends": "node_modules/ocular-dev-tools/templates/.nycrc"
+  "extends": "node_modules/ocular-dev-tools/templates/.nycrc",
+  "include": [
+    "modules/**/src/**/*.js"
+  ]
 }
