+++ package.json
@@ -36,15 +36,15 @@
     "update-release-branch": "scripts/update-release-branch.sh"
   },
   "resolutions": {
-    "lerna": "3.14.1"
+    "puppeteer": "1.12.0"
   },
   "browser": {
     "jsdom": false
   },
   "devDependencies": {
     "@luma.gl/effects": "^7.1.0-beta.1",
-    "@probe.gl/bench": "^3.0.1",
-    "@probe.gl/test-utils": "^3.0.1",
+    "@probe.gl/bench": "^3.0.2",
+    "@probe.gl/test-utils": "^3.0.2",
     "babel-loader": "^8.0.0",
     "babel-plugin-inline-webgl-constants": "^1.0.0",
     "babel-plugin-remove-glsl-comments": "^0.1.0",
@@ -58,7 +58,7 @@
     "gl": "^4.2.2",
     "glsl-transpiler": "^1.8.3",
     "jsdom": "^15.0.0",
-    "ocular-dev-tools": "0.0.18",
+    "ocular-dev-tools": "0.0.27",
     "png.js": "^0.1.1",
     "pre-commit": "^1.2.2",
     "pre-push": "^0.1.1",
