+++ test/bench/node.js
@@ -1,0 +1,3 @@
+require('reify');
+
+require('./index');
