+++ test/browser.js
@@ -43,6 +43,15 @@ enableDOMLogging({
 
 test('deck.gl', t => {
   require('./modules');
+
+  // Tests currently only work in browser
+  require('./modules/json/json-render.spec');
+  require('./modules/main/bundle');
+  require('./modules/aggregation-layers/utils/gpu-grid-aggregator.spec');
+  require('./modules/aggregation-layers/utils/grid-aggregation-utils.spec');
+  require('./modules/core/lib/pick-layers.spec');
+
   require('./render');
+
   t.end();
 });
