+++ test/modules/index.js
@@ -29,31 +29,3 @@ import './mapbox';
 import './json';
 import './react';
 import './jupyter-widget';
-
-if (typeof document !== 'undefined') {
-  // Tests currently only work in browser
-  require('./json/json-render.spec');
-
-  require('./main/bundle');
-  require('./aggregation-layers/utils/gpu-grid-aggregator.spec');
-  require('./aggregation-layers/utils/grid-aggregation-utils.spec');
-  require('./core/lib/pick-layers.spec');
-} else if (typeof global !== 'undefined') {
-  // Running in Node
-  const {JSDOM} = require('jsdom');
-  const dom = new JSDOM(`<!DOCTYPE html>`);
-  // These globals are required by @jupyter-widgets/base
-  /* global global */
-  global.window = dom.window;
-  global.navigator = dom.window.navigator;
-  global.document = dom.window.document;
-  global.Element = dom.window.Element;
-  global.__JSDOM__ = true;
-
-  const {gl} = require('@deck.gl/test-utils');
-  // Create a dummy canvas for the headless gl context
-  const canvas = global.document.createElement('canvas');
-  canvas.width = gl.drawingBufferWidth;
-  canvas.height = gl.drawingBufferHeight;
-  gl.canvas = canvas;
-}
